* misc: the last AROS change added some compiler warnings when compiling for all   the other targets, especially when using format functions like snprintf().   Now the proper type casts are added to avoid these compiler warning   conflicts.

+++ src/BayesFilter.c
@@ -508,7 +508,7 @@ static ULONG countChars(STRPTR str, TEXT c)
 static void tokenizerTokenizeASCIIWord(struct Tokenizer *t,
                                        STRPTR word)
 {
-  ULONG length;
+  size_t length;
 
   ENTER();
