+++ src/YAM_MA.c
@@ -3166,7 +3166,7 @@ void MA_ChangeSubject(struct Mail *mail, char *subj)
       char newfile[SIZE_PATHFILE];
       FILE *newfh;
 
-      snprintf(tfname, sizeof(tfname), "YAMt%08x.tmp", GetUniqueID());
+      snprintf(tfname, sizeof(tfname), "YAMt%08x.tmp", (unsigned int)GetUniqueID());
       AddPath(newfile, GetFolderDir(fo), tfname, sizeof(newfile));
 
       if((newfh = fopen(newfile, "w")) != NULL)
